Adressing stickler comments
